pyproject.toml
[project]
name = "read-later"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiogram>=3.18.0",
    "python-dotenv>=1.0.1",
]
